Extract kanban shared types to break circular import
